application/backend/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.0
google-cloud-firestore==2.20.2
gunicorn==21.2.0